Update Board.java
Fixed indentation from the end of the constructor onwards. This is what was causing 90% of the merge conflict: indentation is part of the line, so GitHub though we'd changed hundreds of lines of code. Please please please always run Correct Indentation on your code by selecting all of it and hitting ctrl+I (that's a capital i) in Eclipse.